chore(test): prepare test files for refactor
