+++ app/ante/cosmos/authz_test.go
@@ -6,9 +6,11 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	sdkvesting "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	"github.com/cosmos/cosmos-sdk/x/authz"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -267,6 +269,163 @@ func TestAuthzLimiterDecorator(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 			}
+		})
+	}
+}
+
+
+func (suite *AnteTestSuite) TestRejectMsgsInAuthz() {
+	testPrivKeys, testAddresses, err := generatePrivKeyAddressPairs(10)
+	suite.Require().NoError(err)
+
+	distantFuture := time.Date(9000, 1, 1, 0, 0, 0, 0, time.UTC)
+
+	newMsgGrant := func(msgTypeUrl string) *authz.MsgGrant {
+		msg, err := authz.NewMsgGrant(
+			testAddresses[0],
+			testAddresses[1],
+			authz.NewGenericAuthorization(msgTypeUrl),
+			&distantFuture,
+		)
+		if err != nil {
+			panic(err)
+		}
+		return msg
+	}
+
+	testcases := []struct {
+		name         string
+		msgs         []sdk.Msg
+		expectedCode uint32
+		isEIP712     bool
+	}{
+		{
+			name:         "a MsgGrant with MsgEthereumTx typeURL on the authorization field is blocked",
+			msgs:         []sdk.Msg{newMsgGrant(sdk.MsgTypeURL(&evmtypes.MsgEthereumTx{}))},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+		{
+			name:         "a MsgGrant with MsgCreateVestingAccount typeURL on the authorization field is blocked",
+			msgs:         []sdk.Msg{newMsgGrant(sdk.MsgTypeURL(&sdkvesting.MsgCreateVestingAccount{}))},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+		{
+			name:         "a MsgGrant with MsgEthereumTx typeURL on the authorization field included on EIP712 tx is blocked",
+			msgs:         []sdk.Msg{newMsgGrant(sdk.MsgTypeURL(&evmtypes.MsgEthereumTx{}))},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+			isEIP712:     true,
+		},
+		{
+			name: "a MsgExec with nested messages (valid: MsgSend and invalid: MsgEthereumTx) is blocked",
+			msgs: []sdk.Msg{
+				newMsgExec(
+					testAddresses[1],
+					[]sdk.Msg{
+						banktypes.NewMsgSend(
+							testAddresses[0],
+							testAddresses[3],
+							sdk.NewCoins(sdk.NewInt64Coin(evmtypes.DefaultEVMDenom, 100e6)),
+						),
+						&evmtypes.MsgEthereumTx{},
+					},
+				),
+			},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+		{
+			name: "a MsgExec with nested MsgExec messages that has invalid messages is blocked",
+			msgs: []sdk.Msg{
+				createNestedMsgExec(
+					testAddresses[1],
+					2,
+					[]sdk.Msg{
+						&evmtypes.MsgEthereumTx{},
+					},
+				),
+			},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+		{
+			name: "a MsgExec with more nested MsgExec messages than allowed and with valid messages is blocked",
+			msgs: []sdk.Msg{
+				createNestedMsgExec(
+					testAddresses[1],
+					6,
+					[]sdk.Msg{
+						banktypes.NewMsgSend(
+							testAddresses[0],
+							testAddresses[3],
+							sdk.NewCoins(sdk.NewInt64Coin(evmtypes.DefaultEVMDenom, 100e6)),
+						),
+					},
+				),
+			},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+		{
+			name: "two MsgExec messages NOT containing a blocked msg but between the two have more nesting than the allowed. Then, is blocked",
+			msgs: []sdk.Msg{
+				createNestedMsgExec(
+					testAddresses[1],
+					5,
+					[]sdk.Msg{
+						banktypes.NewMsgSend(
+							testAddresses[0],
+							testAddresses[3],
+							sdk.NewCoins(sdk.NewInt64Coin(evmtypes.DefaultEVMDenom, 100e6)),
+						),
+					},
+				),
+				createNestedMsgExec(
+					testAddresses[1],
+					5,
+					[]sdk.Msg{
+						banktypes.NewMsgSend(
+							testAddresses[0],
+							testAddresses[3],
+							sdk.NewCoins(sdk.NewInt64Coin(evmtypes.DefaultEVMDenom, 100e6)),
+						),
+					},
+				),
+			},
+			expectedCode: sdkerrors.ErrUnauthorized.ABCICode(),
+		},
+	}
+
+	for _, tc := range testcases {
+		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
+			suite.SetupTest()
+			var (
+				tx  sdk.Tx
+				err error
+			)
+
+			if tc.isEIP712 {
+				tx, err = createEIP712CosmosTx(testAddresses[0], testPrivKeys[0], tc.msgs)
+			} else {
+				tx, err = createTx(testPrivKeys[0], tc.msgs...)
+			}
+			suite.Require().NoError(err)
+
+			txEncoder := suite.clientCtx.TxConfig.TxEncoder()
+			bz, err := txEncoder(tx)
+			suite.Require().NoError(err)
+
+			resCheckTx := suite.app.CheckTx(
+				abci.RequestCheckTx{
+					Tx:   bz,
+					Type: abci.CheckTxType_New,
+				},
+			)
+			suite.Require().Equal(resCheckTx.Code, tc.expectedCode, resCheckTx.Log)
+
+			// TODO uncomment this on v12 release. ATM the anteHandler works on CheckTx mode
+			// resDeliverTx := suite.app.DeliverTx(
+			// 	abci.RequestDeliverTx{
+			// 		Tx: bz,
+			// 	},
+			// )
+			// suite.Require().Equal(resDeliverTx.Code, tc.expectedCode, resDeliverTx.Log)
 		})
 	}
 }
