+++ app/ante/setup_test.go
@@ -1,0 +1,73 @@
+package ante_test
+
+import (
+	"testing"
+	"time"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/evmos/evmos/v11/app"
+	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
+	"github.com/evmos/evmos/v11/utils"
+	feemarkettypes "github.com/evmos/evmos/v11/x/feemarket/types"
+	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
+	"github.com/tendermint/tendermint/crypto/tmhash"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
+	"github.com/tendermint/tendermint/version"
+)
+
+var s *AnteTestSuite
+
+type AnteTestSuite struct {
+	suite.Suite
+
+	ctx   sdk.Context
+	app   *app.Evmos
+	denom string
+}
+
+func (suite *AnteTestSuite) SetupTest() {
+	t := suite.T()
+	privCons, err := ethsecp256k1.GenerateKey()
+	require.NoError(t, err)
+	consAddress := sdk.ConsAddress(privCons.PubKey().Address())
+
+	isCheckTx := false
+	suite.app = app.Setup(isCheckTx, feemarkettypes.DefaultGenesisState())
+	suite.Require().NotNil(suite.app.AppCodec())
+	suite.ctx = suite.app.BaseApp.NewContext(isCheckTx, tmproto.Header{
+		Height:          1,
+		ChainID:         "evmos_9001-1",
+		Time:            time.Now().UTC(),
+		ProposerAddress: consAddress.Bytes(),
+
+		Version: tmversion.Consensus{
+			Block: version.BlockProtocol,
+		},
+		LastBlockId: tmproto.BlockID{
+			Hash: tmhash.Sum([]byte("block_id")),
+			PartSetHeader: tmproto.PartSetHeader{
+				Total: 11,
+				Hash:  tmhash.Sum([]byte("partset_header")),
+			},
+		},
+		AppHash:            tmhash.Sum([]byte("app")),
+		DataHash:           tmhash.Sum([]byte("data")),
+		EvidenceHash:       tmhash.Sum([]byte("evidence")),
+		ValidatorsHash:     tmhash.Sum([]byte("validators")),
+		NextValidatorsHash: tmhash.Sum([]byte("next_validators")),
+		ConsensusHash:      tmhash.Sum([]byte("consensus")),
+		LastResultsHash:    tmhash.Sum([]byte("last_result")),
+	})
+
+	suite.denom = utils.BaseDenom
+	evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+	evmParams.EvmDenom = suite.denom
+	_ = suite.app.EvmKeeper.SetParams(suite.ctx, evmParams)
+}
+
+func TestAnteTestSuite(t *testing.T) {
+	s = new(AnteTestSuite)
+	suite.Run(t, s)
+}
