+++ app/ante/sigverify_test.go
@@ -1,8 +1,6 @@
 package ante_test
 
 import (
-	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -23,33 +21,6 @@ import (
 	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
 	"github.com/evmos/evmos/v11/encoding"
 )
-
-func generatePubKeysAndSignatures(n int, msg []byte, _ bool) (pubkeys []cryptotypes.PubKey, signatures [][]byte) {
-	pubkeys = make([]cryptotypes.PubKey, n)
-	signatures = make([][]byte, n)
-	for i := 0; i < n; i++ {
-		privkey, _ := ethsecp256k1.GenerateKey()
-		pubkeys[i] = privkey.PubKey()
-		signatures[i], _ = privkey.Sign(msg)
-	}
-	return
-}
-
-func expectedGasCostByKeys(pubkeys []cryptotypes.PubKey) uint64 {
-	cost := uint64(0)
-	for _, pubkey := range pubkeys {
-		pubkeyType := strings.ToLower(fmt.Sprintf("%T", pubkey))
-		switch {
-		case strings.Contains(pubkeyType, "ed25519"):
-			cost += authtypes.DefaultSigVerifyCostED25519
-		case strings.Contains(pubkeyType, "ethsecp256k1"):
-			cost += ante.Secp256k1VerifyCost
-		default:
-			panic("unexpected key type")
-		}
-	}
-	return cost
-}
 
 func TestConsumeSignatureVerificationGas(t *testing.T) {
 	params := authtypes.DefaultParams()
