+++ app/ante/utils_test.go
@@ -1,73 +1,38 @@
 package ante_test
 
 import (
-	"testing"
-	"time"
+	"fmt"
+	"strings"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/evmos/evmos/v11/app"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	"github.com/evmos/evmos/v11/app/ante"
 	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
-	"github.com/evmos/evmos/v11/utils"
-	feemarkettypes "github.com/evmos/evmos/v11/x/feemarket/types"
-	"github.com/stretchr/testify/require"
-	"github.com/stretchr/testify/suite"
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
-	"github.com/tendermint/tendermint/version"
 )
 
-var s *AnteTestSuite
-
-type AnteTestSuite struct {
-	suite.Suite
-
-	ctx   sdk.Context
-	app   *app.Evmos
-	denom string
+func generatePubKeysAndSignatures(n int, msg []byte, _ bool) (pubkeys []cryptotypes.PubKey, signatures [][]byte) {
+	pubkeys = make([]cryptotypes.PubKey, n)
+	signatures = make([][]byte, n)
+	for i := 0; i < n; i++ {
+		privkey, _ := ethsecp256k1.GenerateKey()
+		pubkeys[i] = privkey.PubKey()
+		signatures[i], _ = privkey.Sign(msg)
+	}
+	return
 }
 
-func (suite *AnteTestSuite) SetupTest() {
-	t := suite.T()
-	privCons, err := ethsecp256k1.GenerateKey()
-	require.NoError(t, err)
-	consAddress := sdk.ConsAddress(privCons.PubKey().Address())
-
-	isCheckTx := false
-	suite.app = app.Setup(isCheckTx, feemarkettypes.DefaultGenesisState())
-	suite.Require().NotNil(suite.app.AppCodec())
-	suite.ctx = suite.app.BaseApp.NewContext(isCheckTx, tmproto.Header{
-		Height:          1,
-		ChainID:         "evmos_9001-1",
-		Time:            time.Now().UTC(),
-		ProposerAddress: consAddress.Bytes(),
-
-		Version: tmversion.Consensus{
-			Block: version.BlockProtocol,
-		},
-		LastBlockId: tmproto.BlockID{
-			Hash: tmhash.Sum([]byte("block_id")),
-			PartSetHeader: tmproto.PartSetHeader{
-				Total: 11,
-				Hash:  tmhash.Sum([]byte("partset_header")),
-			},
-		},
-		AppHash:            tmhash.Sum([]byte("app")),
-		DataHash:           tmhash.Sum([]byte("data")),
-		EvidenceHash:       tmhash.Sum([]byte("evidence")),
-		ValidatorsHash:     tmhash.Sum([]byte("validators")),
-		NextValidatorsHash: tmhash.Sum([]byte("next_validators")),
-		ConsensusHash:      tmhash.Sum([]byte("consensus")),
-		LastResultsHash:    tmhash.Sum([]byte("last_result")),
-	})
-
-	suite.denom = utils.BaseDenom
-	evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
-	evmParams.EvmDenom = suite.denom
-	_ = suite.app.EvmKeeper.SetParams(suite.ctx, evmParams)
-}
-
-func TestAnteTestSuite(t *testing.T) {
-	s = new(AnteTestSuite)
-	suite.Run(t, s)
+func expectedGasCostByKeys(pubkeys []cryptotypes.PubKey) uint64 {
+	cost := uint64(0)
+	for _, pubkey := range pubkeys {
+		pubkeyType := strings.ToLower(fmt.Sprintf("%T", pubkey))
+		switch {
+		case strings.Contains(pubkeyType, "ed25519"):
+			cost += authtypes.DefaultSigVerifyCostED25519
+		case strings.Contains(pubkeyType, "ethsecp256k1"):
+			cost += ante.Secp256k1VerifyCost
+		default:
+			panic("unexpected key type")
+		}
+	}
+	return cost
 }
