+++ x/claims/keeper/integration_test.go
@@ -1,41 +1,20 @@
 package keeper_test
 
 import (
-	"encoding/json"
 	"math"
-	"math/big"
-	"strconv"
 	"time"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	sdkmath "cosmossdk.io/math"
-	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	abci "github.com/tendermint/tendermint/abci/types"
 
-	"github.com/evmos/evmos/v11/app"
-	"github.com/evmos/evmos/v11/contracts"
 	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
-	"github.com/evmos/evmos/v11/encoding"
-	"github.com/evmos/evmos/v11/server/config"
-	"github.com/evmos/evmos/v11/tests"
 	"github.com/evmos/evmos/v11/testutil"
 	"github.com/evmos/evmos/v11/utils"
 	"github.com/evmos/evmos/v11/x/claims/types"
-	evm "github.com/evmos/evmos/v11/x/evm/types"
-	incentivestypes "github.com/evmos/evmos/v11/x/incentives/types"
 	inflationtypes "github.com/evmos/evmos/v11/x/inflation/types"
 )
 
@@ -333,183 +312,3 @@ var _ = Describe("Claiming", Ordered, func() {
 		})
 	})
 })
-
-func getAddr(priv *ethsecp256k1.PrivKey) sdk.AccAddress {
-	return sdk.AccAddress(priv.PubKey().Address().Bytes())
-}
-
-func delegate(priv *ethsecp256k1.PrivKey, delegateAmount sdk.Coin) {
-	accountAddress := sdk.AccAddress(priv.PubKey().Address().Bytes())
-
-	val, err := sdk.ValAddressFromBech32(s.validator.OperatorAddress)
-	s.Require().NoError(err)
-
-	delegateMsg := stakingtypes.NewMsgDelegate(accountAddress, val, delegateAmount)
-	deliverTx(priv, delegateMsg)
-}
-
-func govProposal(priv *ethsecp256k1.PrivKey) uint64 {
-	stakeDenom := stakingtypes.DefaultParams().BondDenom
-	accountAddress := sdk.AccAddress(priv.PubKey().Address().Bytes())
-	contractAddress := deployContract(priv)
-	content := incentivestypes.NewRegisterIncentiveProposal(
-		"test",
-		"description",
-		contractAddress.String(),
-		sdk.DecCoins{sdk.NewDecCoinFromDec(utils.BaseDenom, sdk.NewDecWithPrec(5, 2))},
-		1000,
-	)
-
-	deposit := sdk.NewCoins(sdk.NewCoin(stakeDenom, sdk.NewInt(100000000)))
-	msg, err := govv1beta1.NewMsgSubmitProposal(content, deposit, accountAddress)
-	s.Require().NoError(err)
-
-	res := deliverTx(priv, msg)
-	submitEvent := res.GetEvents()[8]
-	Expect(submitEvent.Type).To(Equal("submit_proposal"))
-	Expect(string(submitEvent.Attributes[0].Key)).To(Equal("proposal_id"))
-
-	proposalID, err := strconv.ParseUint(string(submitEvent.Attributes[0].Value), 10, 64)
-	s.Require().NoError(err)
-
-	return proposalID
-}
-
-func vote(priv *ethsecp256k1.PrivKey, proposalID uint64) {
-	accountAddress := sdk.AccAddress(priv.PubKey().Address().Bytes())
-
-	voteMsg := govv1beta1.NewMsgVote(accountAddress, proposalID, govv1beta1.OptionAbstain)
-	deliverTx(priv, voteMsg)
-}
-
-func sendEthToSelf(priv *ethsecp256k1.PrivKey) {
-	chainID := s.app.EvmKeeper.ChainID()
-	from := common.BytesToAddress(priv.PubKey().Address().Bytes())
-	nonce := s.app.EvmKeeper.GetNonce(s.ctx, from)
-
-	msgEthereumTx := evm.NewTx(chainID, nonce, &from, nil, 100000, nil, s.app.FeeMarketKeeper.GetBaseFee(s.ctx), big.NewInt(1), nil, &ethtypes.AccessList{})
-	msgEthereumTx.From = from.String()
-	performEthTx(priv, msgEthereumTx)
-}
-
-func deployContract(priv *ethsecp256k1.PrivKey) common.Address {
-	chainID := s.app.EvmKeeper.ChainID()
-	from := common.BytesToAddress(priv.PubKey().Address().Bytes())
-	nonce := s.app.EvmKeeper.GetNonce(s.ctx, from)
-
-	ctorArgs, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("", "Test", "TTT", uint8(18))
-	s.Require().NoError(err)
-
-	data := append(contracts.ERC20MinterBurnerDecimalsContract.Bin, ctorArgs...) //nolint:gocritic
-	args, err := json.Marshal(&evm.TransactionArgs{
-		From: &s.address,
-		Data: (*hexutil.Bytes)(&data),
-	})
-	s.Require().NoError(err)
-
-	ctx := sdk.WrapSDKContext(s.ctx)
-	res, err := s.queryClientEvm.EstimateGas(ctx, &evm.EthCallRequest{
-		Args:   args,
-		GasCap: config.DefaultGasCap,
-	})
-	s.Require().NoError(err)
-
-	msgEthereumTx := evm.NewTxContract(chainID, nonce, nil, res.Gas, nil, s.app.FeeMarketKeeper.GetBaseFee(s.ctx), big.NewInt(1), data, &ethtypes.AccessList{})
-	msgEthereumTx.From = from.String()
-
-	performEthTx(priv, msgEthereumTx)
-	s.Commit()
-
-	contractAddress := crypto.CreateAddress(from, nonce)
-	acc := s.app.EvmKeeper.GetAccountWithoutBalance(s.ctx, contractAddress)
-	s.Require().NotEmpty(acc)
-	s.Require().True(acc.IsContract())
-	return contractAddress
-}
-
-func performEthTx(priv *ethsecp256k1.PrivKey, msgEthereumTx *evm.MsgEthereumTx) {
-	// Sign transaction
-	err := msgEthereumTx.Sign(s.ethSigner, tests.NewSigner(priv))
-	s.Require().NoError(err)
-
-	// Assemble transaction from fields
-	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
-	txBuilder := encodingConfig.TxConfig.NewTxBuilder()
-	tx, err := msgEthereumTx.BuildTx(txBuilder, utils.BaseDenom)
-	s.Require().NoError(err)
-
-	// Encode transaction by default Tx encoder and broadcasted over the network
-	txEncoder := encodingConfig.TxConfig.TxEncoder()
-	bz, err := txEncoder(tx)
-	s.Require().NoError(err)
-
-	req := abci.RequestDeliverTx{Tx: bz}
-	res := s.app.BaseApp.DeliverTx(req)
-	Expect(res.IsOK()).To(Equal(true))
-}
-
-func deliverTx(priv *ethsecp256k1.PrivKey, msgs ...sdk.Msg) abci.ResponseDeliverTx {
-	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
-	accountAddress := sdk.AccAddress(priv.PubKey().Address().Bytes())
-
-	txBuilder := encodingConfig.TxConfig.NewTxBuilder()
-
-	txBuilder.SetGasLimit(1000000)
-	txBuilder.SetFeeAmount(sdk.Coins{defaultTxFee})
-
-	err := txBuilder.SetMsgs(msgs...)
-	s.Require().NoError(err)
-
-	seq, err := s.app.AccountKeeper.GetSequence(s.ctx, accountAddress)
-	s.Require().NoError(err)
-
-	// First round: we gather all the signer infos. We use the "set empty
-	// signature" hack to do that.
-	sigV2 := signing.SignatureV2{
-		PubKey: priv.PubKey(),
-		Data: &signing.SingleSignatureData{
-			SignMode:  encodingConfig.TxConfig.SignModeHandler().DefaultMode(),
-			Signature: nil,
-		},
-		Sequence: seq,
-	}
-
-	sigsV2 := []signing.SignatureV2{sigV2}
-
-	err = txBuilder.SetSignatures(sigsV2...)
-	s.Require().NoError(err)
-
-	// Second round: all signer infos are set, so each signer can sign.
-	accNumber := s.app.AccountKeeper.GetAccount(s.ctx, accountAddress).GetAccountNumber()
-	signerData := authsigning.SignerData{
-		ChainID:       s.ctx.ChainID(),
-		AccountNumber: accNumber,
-		Sequence:      seq,
-	}
-	sigV2, err = tx.SignWithPrivKey(
-		encodingConfig.TxConfig.SignModeHandler().DefaultMode(), signerData,
-		txBuilder, priv, encodingConfig.TxConfig,
-		seq,
-	)
-	s.Require().NoError(err)
-
-	sigsV2 = []signing.SignatureV2{sigV2}
-	err = txBuilder.SetSignatures(sigsV2...)
-	s.Require().NoError(err)
-
-	// bz are bytes to be broadcasted over the network
-	bz, err := encodingConfig.TxConfig.TxEncoder()(txBuilder.GetTx())
-	s.Require().NoError(err)
-
-	req := abci.RequestDeliverTx{Tx: bz}
-	res := s.app.BaseApp.DeliverTx(req)
-	Expect(res.IsOK()).To(Equal(true), res.Log)
-	return res
-}
-
-func getEthTxFee() sdk.Coin {
-	baseFee := s.app.FeeMarketKeeper.GetBaseFee(s.ctx)
-	baseFee.Mul(baseFee, big.NewInt(100000))
-	feeAmt := baseFee.Quo(baseFee, big.NewInt(2))
-	return sdk.NewCoin(utils.BaseDenom, sdkmath.NewIntFromBigInt(feeAmt))
-}
