+++ x/erc20/keeper/erc20_test.go
@@ -1,0 +1,75 @@
+package keeper_test
+
+import (
+	"fmt"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/evmos/evmos/v11/contracts"
+	"github.com/evmos/evmos/v11/x/erc20/types"
+	evm "github.com/evmos/evmos/v11/x/evm/types"
+)
+
+func (suite *KeeperTestSuite) MintERC20Token(contractAddr, from, to common.Address, amount *big.Int) *evm.MsgEthereumTx {
+	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("mint", to, amount)
+	suite.Require().NoError(err)
+	return suite.sendTx(contractAddr, from, transferData)
+}
+
+func (suite *KeeperTestSuite) TransferERC20TokenToModule(contractAddr, from common.Address, amount *big.Int) *evm.MsgEthereumTx {
+	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("transfer", types.ModuleAddress, amount)
+	suite.Require().NoError(err)
+	return suite.sendTx(contractAddr, from, transferData)
+}
+
+func (suite *KeeperTestSuite) GrantERC20Token(contractAddr, from, to common.Address, roleString string) *evm.MsgEthereumTx {
+	// 0xCc508cD0818C85b8b8a1aB4cEEef8d981c8956A6 MINTER_ROLE
+	role := crypto.Keccak256([]byte(roleString))
+	// needs to be an array not a slice
+	var v [32]byte
+	copy(v[:], role)
+
+	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("grantRole", v, to)
+	suite.Require().NoError(err)
+	return suite.sendTx(contractAddr, from, transferData)
+}
+
+func (suite *KeeperTestSuite) BalanceOf(contract, account common.Address) interface{} {
+	erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
+
+	res, err := suite.app.Erc20Keeper.CallEVM(suite.ctx, erc20, types.ModuleAddress, contract, false, "balanceOf", account)
+	if err != nil {
+		return nil
+	}
+
+	unpacked, err := erc20.Unpack("balanceOf", res.Ret)
+	if err != nil {
+		return nil
+	}
+	if len(unpacked) == 0 {
+		return nil
+	}
+
+	return unpacked[0]
+}
+
+func (suite *KeeperTestSuite) NameOf(contract common.Address) string {
+	erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
+
+	res, err := suite.app.Erc20Keeper.CallEVM(suite.ctx, erc20, types.ModuleAddress, contract, false, "name")
+	suite.Require().NoError(err)
+	suite.Require().NotNil(res)
+
+	unpacked, err := erc20.Unpack("name", res.Ret)
+	suite.Require().NoError(err)
+	suite.Require().NotEmpty(unpacked)
+
+	return fmt.Sprintf("%v", unpacked[0])
+}
+
+func (suite *KeeperTestSuite) TransferERC20Token(contractAddr, from, to common.Address, amount *big.Int) *evm.MsgEthereumTx {
+	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("transfer", to, amount)
+	suite.Require().NoError(err)
+	return suite.sendTx(contractAddr, from, transferData)
+}
