+++ x/erc20/keeper/utils_test.go
@@ -1,115 +1,65 @@
 package keeper_test
 
 import (
-	"context"
 	"encoding/json"
-	"fmt"
 	"math/big"
 	"strconv"
-	"testing"
 	"time"
 
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-
 	"github.com/cosmos/cosmos-sdk/baseapp"
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-	"github.com/stretchr/testify/suite"
-	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/crypto/tmhash"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
-	"github.com/tendermint/tendermint/version"
-
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/core/vm"
-	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
-	"github.com/evmos/evmos/v11/server/config"
-	"github.com/evmos/evmos/v11/tests"
-	"github.com/evmos/evmos/v11/x/evm/statedb"
-	evm "github.com/evmos/evmos/v11/x/evm/types"
-	feemarkettypes "github.com/evmos/evmos/v11/x/feemarket/types"
-
-	"github.com/evmos/evmos/v11/app"
-	"github.com/evmos/evmos/v11/contracts"
-	ibctesting "github.com/evmos/evmos/v11/ibc/testing"
-	claimstypes "github.com/evmos/evmos/v11/x/claims/types"
-	"github.com/evmos/evmos/v11/x/erc20/types"
-	inflationtypes "github.com/evmos/evmos/v11/x/inflation/types"
-
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
 	clienttypes "github.com/cosmos/ibc-go/v6/modules/core/02-client/types"
 	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
 	ibcgotesting "github.com/cosmos/ibc-go/v6/testing"
 	ibcgotestinghelpers "github.com/cosmos/ibc-go/v6/testing/simapp/helpers"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/evmos/evmos/v11/app"
+	"github.com/evmos/evmos/v11/contracts"
+	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
+	ibctesting "github.com/evmos/evmos/v11/ibc/testing"
+	"github.com/evmos/evmos/v11/server/config"
+	"github.com/evmos/evmos/v11/tests"
 	teststypes "github.com/evmos/evmos/v11/types/tests"
 	"github.com/evmos/evmos/v11/utils"
+	claimstypes "github.com/evmos/evmos/v11/x/claims/types"
+	"github.com/evmos/evmos/v11/x/erc20/types"
+	"github.com/evmos/evmos/v11/x/evm/statedb"
+	evm "github.com/evmos/evmos/v11/x/evm/types"
+	feemarkettypes "github.com/evmos/evmos/v11/x/feemarket/types"
+	inflationtypes "github.com/evmos/evmos/v11/x/inflation/types"
+	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/crypto/tmhash"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
+	"github.com/tendermint/tendermint/version"
 )
 
-type KeeperTestSuite struct {
-	suite.Suite
-
-	ctx              sdk.Context
-	app              *app.Evmos
-	queryClientEvm   evm.QueryClient
-	queryClient      types.QueryClient
-	address          common.Address
-	consAddress      sdk.ConsAddress
-	clientCtx        client.Context //nolint:unused
-	ethSigner        ethtypes.Signer
-	validator        stakingtypes.Validator
-	signer           keyring.Signer
-	mintFeeCollector bool
-
-	coordinator *ibcgotesting.Coordinator
-
-	// testing chains used for convenience and readability
-	EvmosChain      *ibcgotesting.TestChain
-	IBCOsmosisChain *ibcgotesting.TestChain
-	IBCCosmosChain  *ibcgotesting.TestChain
-
-	pathOsmosisEvmos  *ibctesting.Path
-	pathCosmosEvmos   *ibctesting.Path
-	pathOsmosisCosmos *ibctesting.Path
-
-	suiteIBCTesting bool
-}
-
-var (
-	s *KeeperTestSuite
-	// sendAndReceiveMsgFee corresponds to the fees paid on Evmos chain when calling the SendAndReceive function
-	// This function makes 3 cosmos txs under the hood
-	sendAndReceiveMsgFee = sdk.NewInt(ibctesting.DefaultFeeAmt * 3)
-	// sendBackCoinsFee corresponds to the fees paid on Evmos chain when calling the SendBackCoins function
-	// or calling the SendAndReceive from another chain to Evmos
-	// This function makes 2 cosmos txs under the hood
-	sendBackCoinsFee = sdk.NewInt(ibctesting.DefaultFeeAmt * 2)
-)
-
-func TestKeeperTestSuite(t *testing.T) {
-	s = new(KeeperTestSuite)
-	suite.Run(t, s)
-
-	// Run Ginkgo integration tests
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Keeper Suite")
-}
-
-func (suite *KeeperTestSuite) SetupTest() {
-	suite.DoSetupTest(suite.T())
+func CreatePacket(amount, denom, sender, receiver, srcPort, srcChannel, dstPort, dstChannel string, seq, timeout uint64) channeltypes.Packet {
+	transfer := transfertypes.FungibleTokenPacketData{
+		Amount:   amount,
+		Denom:    denom,
+		Receiver: sender,
+		Sender:   receiver,
+	}
+	return channeltypes.NewPacket(
+		transfer.GetBytes(),
+		seq,
+		srcPort,
+		srcChannel,
+		dstPort,
+		dstChannel,
+		clienttypes.ZeroHeight(), // timeout height disabled
+		timeout,
+	)
 }
 
 func (suite *KeeperTestSuite) DoSetupTest(t require.TestingT) {
@@ -195,66 +145,6 @@ func (suite *KeeperTestSuite) DoSetupTest(t require.TestingT) {
 	if suite.suiteIBCTesting {
 		suite.SetupIBCTest()
 	}
-}
-
-var timeoutHeight = clienttypes.NewHeight(1000, 1000)
-
-func (suite *KeeperTestSuite) sendAndReceiveMessage(
-	path *ibctesting.Path,
-	originEndpoint *ibctesting.Endpoint,
-	destEndpoint *ibctesting.Endpoint,
-	originChain *ibcgotesting.TestChain,
-	coin string,
-	amount int64,
-	sender, receiver string,
-	seq uint64,
-	ibcCoinMetadata string,
-) {
-	transferMsg := transfertypes.NewMsgTransfer(originEndpoint.ChannelConfig.PortID, originEndpoint.ChannelID, sdk.NewCoin(coin, sdk.NewInt(amount)), sender, receiver, timeoutHeight, 0, "")
-	_, err := ibctesting.SendMsgs(originChain, ibctesting.DefaultFeeAmt, transferMsg)
-	suite.Require().NoError(err) // message committed
-	// Recreate the packet that was sent
-	var transfer transfertypes.FungibleTokenPacketData
-	if ibcCoinMetadata == "" {
-		transfer = transfertypes.NewFungibleTokenPacketData(coin, strconv.Itoa(int(amount)), sender, receiver, "")
-	} else {
-		transfer = transfertypes.NewFungibleTokenPacketData(ibcCoinMetadata, strconv.Itoa(int(amount)), sender, receiver, "")
-	}
-	packet := channeltypes.NewPacket(transfer.GetBytes(), seq, originEndpoint.ChannelConfig.PortID, originEndpoint.ChannelID, destEndpoint.ChannelConfig.PortID, destEndpoint.ChannelID, timeoutHeight, 0)
-	// Receive message on the counterparty side, and send ack
-	err = path.RelayPacket(packet)
-	suite.Require().NoError(err)
-}
-
-func (suite *KeeperTestSuite) SendAndReceiveMessage(path *ibctesting.Path, origin *ibcgotesting.TestChain, coin string, amount int64, sender, receiver string, seq uint64, ibcCoinMetadata string) {
-	// Send coin from A to B
-	suite.sendAndReceiveMessage(path, path.EndpointA, path.EndpointB, origin, coin, amount, sender, receiver, seq, ibcCoinMetadata)
-}
-
-// Send back coins (from path endpoint B to A). In case of IBC coins need to provide ibcCoinMetadata (<port>/<channel>/<denom>, e.g.: "transfer/channel-0/aevmos") as input parameter.
-// We need this to instantiate properly a FungibleTokenPacketData https://github.com/cosmos/ibc-go/blob/main/docs/architecture/adr-001-coin-source-tracing.md
-func (suite *KeeperTestSuite) SendBackCoins(path *ibctesting.Path, origin *ibcgotesting.TestChain, coin string, amount int64, sender, receiver string, seq uint64, ibcCoinMetadata string) {
-	// Send coin from B to A
-	suite.sendAndReceiveMessage(path, path.EndpointB, path.EndpointA, origin, coin, amount, sender, receiver, seq, ibcCoinMetadata)
-}
-
-func CreatePacket(amount, denom, sender, receiver, srcPort, srcChannel, dstPort, dstChannel string, seq, timeout uint64) channeltypes.Packet {
-	transfer := transfertypes.FungibleTokenPacketData{
-		Amount:   amount,
-		Denom:    denom,
-		Receiver: sender,
-		Sender:   receiver,
-	}
-	return channeltypes.NewPacket(
-		transfer.GetBytes(),
-		seq,
-		srcPort,
-		srcChannel,
-		dstPort,
-		dstChannel,
-		clienttypes.ZeroHeight(), // timeout height disabled
-		timeout,
-	)
 }
 
 func (suite *KeeperTestSuite) SetupIBCTest() {
@@ -365,6 +255,8 @@ func (suite *KeeperTestSuite) SetupIBCTest() {
 	suite.Require().NoError(err)
 }
 
+var timeoutHeight = clienttypes.NewHeight(1000, 1000)
+
 func (suite *KeeperTestSuite) StateDB() *statedb.StateDB {
 	return statedb.New(suite.ctx, suite.app.EvmKeeper, statedb.NewEmptyTxConfig(common.BytesToHash(suite.ctx.HeaderHash().Bytes())))
 }
@@ -374,6 +266,75 @@ func (suite *KeeperTestSuite) MintFeeCollector(coins sdk.Coins) {
 	suite.Require().NoError(err)
 	err = suite.app.BankKeeper.SendCoinsFromModuleToModule(suite.ctx, types.ModuleName, authtypes.FeeCollectorName, coins)
 	suite.Require().NoError(err)
+}
+
+func (suite *KeeperTestSuite) sendTx(contractAddr, from common.Address, transferData []byte) *evm.MsgEthereumTx {
+	ctx := sdk.WrapSDKContext(suite.ctx)
+	chainID := suite.app.EvmKeeper.ChainID()
+
+	args, err := json.Marshal(&evm.TransactionArgs{To: &contractAddr, From: &from, Data: (*hexutil.Bytes)(&transferData)})
+	suite.Require().NoError(err)
+	res, err := suite.queryClientEvm.EstimateGas(ctx, &evm.EthCallRequest{
+		Args:   args,
+		GasCap: config.DefaultGasCap,
+	})
+	suite.Require().NoError(err)
+
+	nonce := suite.app.EvmKeeper.GetNonce(suite.ctx, suite.address)
+
+	// Mint the max gas to the FeeCollector to ensure balance in case of refund
+	evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+	suite.MintFeeCollector(sdk.NewCoins(sdk.NewCoin(evmParams.EvmDenom, sdk.NewInt(suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx).Int64()*int64(res.Gas)))))
+
+	ercTransferTx := evm.NewTx(
+		chainID,
+		nonce,
+		&contractAddr,
+		nil,
+		res.Gas,
+		nil,
+		suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx),
+		big.NewInt(1),
+		transferData,
+		&ethtypes.AccessList{}, // accesses
+	)
+
+	ercTransferTx.From = suite.address.Hex()
+	err = ercTransferTx.Sign(ethtypes.LatestSignerForChainID(chainID), suite.signer)
+	suite.Require().NoError(err)
+	rsp, err := suite.app.EvmKeeper.EthereumTx(ctx, ercTransferTx)
+	suite.Require().NoError(err)
+	suite.Require().Empty(rsp.VmError)
+	return ercTransferTx
+}
+
+// Commit commits and starts a new block with an updated context.
+func (suite *KeeperTestSuite) Commit() {
+	suite.CommitAndBeginBlockAfter(time.Hour * 1)
+}
+
+// Commit commits a block at a given time. Reminder: At the end of each
+// Tendermint Consensus round the following methods are run
+//  1. BeginBlock
+//  2. DeliverTx
+//  3. EndBlock
+//  4. Commit
+func (suite *KeeperTestSuite) CommitAndBeginBlockAfter(t time.Duration) {
+	header := suite.ctx.BlockHeader()
+	_ = suite.app.Commit()
+
+	header.Height++
+	header.Time = header.Time.Add(t)
+	suite.app.BeginBlock(abci.RequestBeginBlock{
+		Header: header,
+	})
+
+	// update ctx
+	suite.ctx = suite.app.BaseApp.NewContext(false, header)
+
+	queryHelper := baseapp.NewQueryServerTestHelper(suite.ctx, suite.app.InterfaceRegistry())
+	evm.RegisterQueryServer(queryHelper, suite.app.EvmKeeper)
+	suite.queryClientEvm = evm.NewQueryClient(queryHelper)
 }
 
 // DeployContract deploys the ERC20MinterBurnerDecimalsContract.
@@ -518,228 +479,6 @@ func (suite *KeeperTestSuite) DeployContractDirectBalanceManipulation(name strin
 	return crypto.CreateAddress(suite.address, nonce)
 }
 
-// Commit commits and starts a new block with an updated context.
-func (suite *KeeperTestSuite) Commit() {
-	suite.CommitAndBeginBlockAfter(time.Hour * 1)
-}
-
-// Commit commits a block at a given time. Reminder: At the end of each
-// Tendermint Consensus round the following methods are run
-//  1. BeginBlock
-//  2. DeliverTx
-//  3. EndBlock
-//  4. Commit
-func (suite *KeeperTestSuite) CommitAndBeginBlockAfter(t time.Duration) {
-	header := suite.ctx.BlockHeader()
-	_ = suite.app.Commit()
-
-	header.Height++
-	header.Time = header.Time.Add(t)
-	suite.app.BeginBlock(abci.RequestBeginBlock{
-		Header: header,
-	})
-
-	// update ctx
-	suite.ctx = suite.app.BaseApp.NewContext(false, header)
-
-	queryHelper := baseapp.NewQueryServerTestHelper(suite.ctx, suite.app.InterfaceRegistry())
-	evm.RegisterQueryServer(queryHelper, suite.app.EvmKeeper)
-	suite.queryClientEvm = evm.NewQueryClient(queryHelper)
-}
-
-func (suite *KeeperTestSuite) MintERC20Token(contractAddr, from, to common.Address, amount *big.Int) *evm.MsgEthereumTx {
-	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("mint", to, amount)
-	suite.Require().NoError(err)
-	return suite.sendTx(contractAddr, from, transferData)
-}
-
-func (suite *KeeperTestSuite) TransferERC20TokenToModule(contractAddr, from common.Address, amount *big.Int) *evm.MsgEthereumTx {
-	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("transfer", types.ModuleAddress, amount)
-	suite.Require().NoError(err)
-	return suite.sendTx(contractAddr, from, transferData)
-}
-
-func (suite *KeeperTestSuite) GrantERC20Token(contractAddr, from, to common.Address, roleString string) *evm.MsgEthereumTx {
-	// 0xCc508cD0818C85b8b8a1aB4cEEef8d981c8956A6 MINTER_ROLE
-	role := crypto.Keccak256([]byte(roleString))
-	// needs to be an array not a slice
-	var v [32]byte
-	copy(v[:], role)
-
-	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("grantRole", v, to)
-	suite.Require().NoError(err)
-	return suite.sendTx(contractAddr, from, transferData)
-}
-
-func (suite *KeeperTestSuite) sendTx(contractAddr, from common.Address, transferData []byte) *evm.MsgEthereumTx {
-	ctx := sdk.WrapSDKContext(suite.ctx)
-	chainID := suite.app.EvmKeeper.ChainID()
-
-	args, err := json.Marshal(&evm.TransactionArgs{To: &contractAddr, From: &from, Data: (*hexutil.Bytes)(&transferData)})
-	suite.Require().NoError(err)
-	res, err := suite.queryClientEvm.EstimateGas(ctx, &evm.EthCallRequest{
-		Args:   args,
-		GasCap: config.DefaultGasCap,
-	})
-	suite.Require().NoError(err)
-
-	nonce := suite.app.EvmKeeper.GetNonce(suite.ctx, suite.address)
-
-	// Mint the max gas to the FeeCollector to ensure balance in case of refund
-	evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
-	suite.MintFeeCollector(sdk.NewCoins(sdk.NewCoin(evmParams.EvmDenom, sdk.NewInt(suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx).Int64()*int64(res.Gas)))))
-
-	ercTransferTx := evm.NewTx(
-		chainID,
-		nonce,
-		&contractAddr,
-		nil,
-		res.Gas,
-		nil,
-		suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx),
-		big.NewInt(1),
-		transferData,
-		&ethtypes.AccessList{}, // accesses
-	)
-
-	ercTransferTx.From = suite.address.Hex()
-	err = ercTransferTx.Sign(ethtypes.LatestSignerForChainID(chainID), suite.signer)
-	suite.Require().NoError(err)
-	rsp, err := suite.app.EvmKeeper.EthereumTx(ctx, ercTransferTx)
-	suite.Require().NoError(err)
-	suite.Require().Empty(rsp.VmError)
-	return ercTransferTx
-}
-
-func (suite *KeeperTestSuite) BalanceOf(contract, account common.Address) interface{} {
-	erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
-
-	res, err := suite.app.Erc20Keeper.CallEVM(suite.ctx, erc20, types.ModuleAddress, contract, false, "balanceOf", account)
-	if err != nil {
-		return nil
-	}
-
-	unpacked, err := erc20.Unpack("balanceOf", res.Ret)
-	if err != nil {
-		return nil
-	}
-	if len(unpacked) == 0 {
-		return nil
-	}
-
-	return unpacked[0]
-}
-
-func (suite *KeeperTestSuite) NameOf(contract common.Address) string {
-	erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
-
-	res, err := suite.app.Erc20Keeper.CallEVM(suite.ctx, erc20, types.ModuleAddress, contract, false, "name")
-	suite.Require().NoError(err)
-	suite.Require().NotNil(res)
-
-	unpacked, err := erc20.Unpack("name", res.Ret)
-	suite.Require().NoError(err)
-	suite.Require().NotEmpty(unpacked)
-
-	return fmt.Sprintf("%v", unpacked[0])
-}
-
-func (suite *KeeperTestSuite) TransferERC20Token(contractAddr, from, to common.Address, amount *big.Int) *evm.MsgEthereumTx {
-	transferData, err := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("transfer", to, amount)
-	suite.Require().NoError(err)
-	return suite.sendTx(contractAddr, from, transferData)
-}
-
-var _ types.EVMKeeper = &MockEVMKeeper{}
-
-type MockEVMKeeper struct {
-	mock.Mock
-}
-
-func (m *MockEVMKeeper) GetParams(ctx sdk.Context) evm.Params {
-	args := m.Called(mock.Anything)
-	return args.Get(0).(evm.Params)
-}
-
-func (m *MockEVMKeeper) GetAccountWithoutBalance(ctx sdk.Context, addr common.Address) *statedb.Account {
-	args := m.Called(mock.Anything, mock.Anything)
-	if args.Get(0) == nil {
-		return nil
-	}
-	return args.Get(0).(*statedb.Account)
-}
-
-func (m *MockEVMKeeper) EstimateGas(c context.Context, req *evm.EthCallRequest) (*evm.EstimateGasResponse, error) {
-	args := m.Called(mock.Anything, mock.Anything)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*evm.EstimateGasResponse), args.Error(1)
-}
-
-func (m *MockEVMKeeper) ApplyMessage(ctx sdk.Context, msg core.Message, tracer vm.EVMLogger, commit bool) (*evm.MsgEthereumTxResponse, error) {
-	args := m.Called(mock.Anything, mock.Anything, mock.Anything, mock.Anything)
-
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*evm.MsgEthereumTxResponse), args.Error(1)
-}
-
-var _ types.BankKeeper = &MockBankKeeper{}
-
-type MockBankKeeper struct {
-	mock.Mock
-}
-
-func (b *MockBankKeeper) SendCoinsFromModuleToAccount(ctx sdk.Context, senderModule string, recipientAddr sdk.AccAddress, amt sdk.Coins) error {
-	args := b.Called(mock.Anything, mock.Anything, mock.Anything, mock.Anything)
-	return args.Error(0)
-}
-
-func (b *MockBankKeeper) SendCoinsFromAccountToModule(ctx sdk.Context, senderAddr sdk.AccAddress, recipientModule string, amt sdk.Coins) error {
-	args := b.Called(mock.Anything, mock.Anything, mock.Anything, mock.Anything)
-	return args.Error(0)
-}
-
-func (b *MockBankKeeper) MintCoins(ctx sdk.Context, moduleName string, amt sdk.Coins) error {
-	args := b.Called(mock.Anything, mock.Anything, mock.Anything)
-	return args.Error(0)
-}
-
-func (b *MockBankKeeper) BurnCoins(ctx sdk.Context, moduleName string, amt sdk.Coins) error {
-	args := b.Called(mock.Anything, mock.Anything, mock.Anything)
-	return args.Error(0)
-}
-
-func (b *MockBankKeeper) IsSendEnabledCoin(ctx sdk.Context, coin sdk.Coin) bool {
-	args := b.Called(mock.Anything, mock.Anything)
-	return args.Bool(0)
-}
-
-func (b *MockBankKeeper) BlockedAddr(addr sdk.AccAddress) bool {
-	args := b.Called(mock.Anything)
-	return args.Bool(0)
-}
-
-func (b *MockBankKeeper) GetDenomMetaData(ctx sdk.Context, denom string) (banktypes.Metadata, bool) {
-	args := b.Called(mock.Anything, mock.Anything)
-	return args.Get(0).(banktypes.Metadata), args.Bool(1)
-}
-
-func (b *MockBankKeeper) SetDenomMetaData(ctx sdk.Context, denomMetaData banktypes.Metadata) {
-}
-
-func (b *MockBankKeeper) HasSupply(ctx sdk.Context, denom string) bool {
-	args := b.Called(mock.Anything, mock.Anything)
-	return args.Bool(0)
-}
-
-func (b *MockBankKeeper) GetBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin {
-	args := b.Called(mock.Anything, mock.Anything)
-	return args.Get(0).(sdk.Coin)
-}
-
 // DeployContract deploys the ERC20MinterBurnerDecimalsContract.
 func (suite *KeeperTestSuite) DeployContractToChain(name, symbol string, decimals uint8) (common.Address, error) {
 	ctx := sdk.WrapSDKContext(s.EvmosChain.GetContext())
@@ -780,4 +519,43 @@ func (suite *KeeperTestSuite) DeployContractToChain(name, symbol string, decimal
 
 	suite.Require().Empty(rsp.VmError)
 	return crypto.CreateAddress(from, nonce), nil
+}
+
+func (suite *KeeperTestSuite) sendAndReceiveMessage(
+	path *ibctesting.Path,
+	originEndpoint *ibctesting.Endpoint,
+	destEndpoint *ibctesting.Endpoint,
+	originChain *ibcgotesting.TestChain,
+	coin string,
+	amount int64,
+	sender, receiver string,
+	seq uint64,
+	ibcCoinMetadata string,
+) {
+	transferMsg := transfertypes.NewMsgTransfer(originEndpoint.ChannelConfig.PortID, originEndpoint.ChannelID, sdk.NewCoin(coin, sdk.NewInt(amount)), sender, receiver, timeoutHeight, 0, "")
+	_, err := ibctesting.SendMsgs(originChain, ibctesting.DefaultFeeAmt, transferMsg)
+	suite.Require().NoError(err) // message committed
+	// Recreate the packet that was sent
+	var transfer transfertypes.FungibleTokenPacketData
+	if ibcCoinMetadata == "" {
+		transfer = transfertypes.NewFungibleTokenPacketData(coin, strconv.Itoa(int(amount)), sender, receiver, "")
+	} else {
+		transfer = transfertypes.NewFungibleTokenPacketData(ibcCoinMetadata, strconv.Itoa(int(amount)), sender, receiver, "")
+	}
+	packet := channeltypes.NewPacket(transfer.GetBytes(), seq, originEndpoint.ChannelConfig.PortID, originEndpoint.ChannelID, destEndpoint.ChannelConfig.PortID, destEndpoint.ChannelID, timeoutHeight, 0)
+	// Receive message on the counterparty side, and send ack
+	err = path.RelayPacket(packet)
+	suite.Require().NoError(err)
+}
+
+func (suite *KeeperTestSuite) SendAndReceiveMessage(path *ibctesting.Path, origin *ibcgotesting.TestChain, coin string, amount int64, sender, receiver string, seq uint64, ibcCoinMetadata string) {
+	// Send coin from A to B
+	suite.sendAndReceiveMessage(path, path.EndpointA, path.EndpointB, origin, coin, amount, sender, receiver, seq, ibcCoinMetadata)
+}
+
+// Send back coins (from path endpoint B to A). In case of IBC coins need to provide ibcCoinMetadata (<port>/<channel>/<denom>, e.g.: "transfer/channel-0/aevmos") as input parameter.
+// We need this to instantiate properly a FungibleTokenPacketData https://github.com/cosmos/ibc-go/blob/main/docs/architecture/adr-001-coin-source-tracing.md
+func (suite *KeeperTestSuite) SendBackCoins(path *ibctesting.Path, origin *ibcgotesting.TestChain, coin string, amount int64, sender, receiver string, seq uint64, ibcCoinMetadata string) {
+	// Send coin from B to A
+	suite.sendAndReceiveMessage(path, path.EndpointB, path.EndpointA, origin, coin, amount, sender, receiver, seq, ibcCoinMetadata)
 }
