+++ x/evm/keeper/utils_test.go
@@ -1,517 +1,214 @@
 package keeper_test
 
 import (
+	"encoding/json"
 	"math/big"
 
-	sdkmath "cosmossdk.io/math"
+	"github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	ethparams "github.com/ethereum/go-ethereum/params"
-	"github.com/evmos/evmos/v11/x/evm/keeper"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/evmos/evmos/v11/server/config"
+	"github.com/evmos/evmos/v11/x/evm/statedb"
 	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
+	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
 )
 
-func (suite *KeeperTestSuite) TestCheckSenderBalance() {
-	hundredInt := sdkmath.NewInt(100)
-	zeroInt := sdk.ZeroInt()
-	oneInt := sdk.OneInt()
-	fiveInt := sdkmath.NewInt(5)
-	fiftyInt := sdkmath.NewInt(50)
-	negInt := sdkmath.NewInt(-10)
-
-	testCases := []struct {
-		name            string
-		to              string
-		gasLimit        uint64
-		gasPrice        *sdkmath.Int
-		gasFeeCap       *big.Int
-		gasTipCap       *big.Int
-		cost            *sdkmath.Int
-		from            string
-		accessList      *ethtypes.AccessList
-		expectPass      bool
-		enableFeemarket bool
-	}{
-		{
-			name:       "Enough balance",
-			to:         suite.address.String(),
-			gasLimit:   10,
-			gasPrice:   &oneInt,
-			cost:       &oneInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: true,
-		},
-		{
-			name:       "Equal balance",
-			to:         suite.address.String(),
-			gasLimit:   99,
-			gasPrice:   &oneInt,
-			cost:       &oneInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: true,
-		},
-		{
-			name:       "negative cost",
-			to:         suite.address.String(),
-			gasLimit:   1,
-			gasPrice:   &oneInt,
-			cost:       &negInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: false,
-		},
-		{
-			name:       "Higher gas limit, not enough balance",
-			to:         suite.address.String(),
-			gasLimit:   100,
-			gasPrice:   &oneInt,
-			cost:       &oneInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: false,
-		},
-		{
-			name:       "Higher gas price, enough balance",
-			to:         suite.address.String(),
-			gasLimit:   10,
-			gasPrice:   &fiveInt,
-			cost:       &oneInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: true,
-		},
-		{
-			name:       "Higher gas price, not enough balance",
-			to:         suite.address.String(),
-			gasLimit:   20,
-			gasPrice:   &fiveInt,
-			cost:       &oneInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: false,
-		},
-		{
-			name:       "Higher cost, enough balance",
-			to:         suite.address.String(),
-			gasLimit:   10,
-			gasPrice:   &fiveInt,
-			cost:       &fiftyInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: true,
-		},
-		{
-			name:       "Higher cost, not enough balance",
-			to:         suite.address.String(),
-			gasLimit:   10,
-			gasPrice:   &fiveInt,
-			cost:       &hundredInt,
-			from:       suite.address.String(),
-			accessList: &ethtypes.AccessList{},
-			expectPass: false,
-		},
-		{
-			name:            "Enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        10,
-			gasFeeCap:       big.NewInt(1),
-			cost:            &oneInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      true,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Equal balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        99,
-			gasFeeCap:       big.NewInt(1),
-			cost:            &oneInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      true,
-			enableFeemarket: true,
-		},
-		{
-			name:            "negative cost w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        1,
-			gasFeeCap:       big.NewInt(1),
-			cost:            &negInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      false,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Higher gas limit, not enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        100,
-			gasFeeCap:       big.NewInt(1),
-			cost:            &oneInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      false,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Higher gas price, enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        10,
-			gasFeeCap:       big.NewInt(5),
-			cost:            &oneInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      true,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Higher gas price, not enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        20,
-			gasFeeCap:       big.NewInt(5),
-			cost:            &oneInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      false,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Higher cost, enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        10,
-			gasFeeCap:       big.NewInt(5),
-			cost:            &fiftyInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      true,
-			enableFeemarket: true,
-		},
-		{
-			name:            "Higher cost, not enough balance w/ enableFeemarket",
-			to:              suite.address.String(),
-			gasLimit:        10,
-			gasFeeCap:       big.NewInt(5),
-			cost:            &hundredInt,
-			from:            suite.address.String(),
-			accessList:      &ethtypes.AccessList{},
-			expectPass:      false,
-			enableFeemarket: true,
-		},
-	}
-
-	vmdb := suite.StateDB()
-	vmdb.AddBalance(suite.address, hundredInt.BigInt())
-	balance := vmdb.GetBalance(suite.address)
-	suite.Require().Equal(balance, hundredInt.BigInt())
-	err := vmdb.Commit()
-	suite.Require().NoError(err, "Unexpected error while committing to vmdb: %d", err)
-
-	for i, tc := range testCases {
-		suite.Run(tc.name, func() {
-			to := common.HexToAddress(tc.from)
-
-			var amount, gasPrice, gasFeeCap, gasTipCap *big.Int
-			if tc.cost != nil {
-				amount = tc.cost.BigInt()
-			}
-
-			if tc.enableFeemarket {
-				gasFeeCap = tc.gasFeeCap
-				if tc.gasTipCap == nil {
-					gasTipCap = oneInt.BigInt()
-				} else {
-					gasTipCap = tc.gasTipCap
-				}
-			} else if tc.gasPrice != nil {
-				gasPrice = tc.gasPrice.BigInt()
-			}
-
-			tx := evmtypes.NewTx(zeroInt.BigInt(), 1, &to, amount, tc.gasLimit, gasPrice, gasFeeCap, gasTipCap, nil, tc.accessList)
-			tx.From = tc.from
-
-			txData, _ := evmtypes.UnpackTxData(tx.Data)
-
-			acct := suite.app.EvmKeeper.GetAccountOrEmpty(suite.ctx, suite.address)
-			err := keeper.CheckSenderBalance(
-				sdkmath.NewIntFromBigInt(acct.Balance),
-				txData,
-			)
-
-			if tc.expectPass {
-				suite.Require().NoError(err, "valid test %d failed", i)
-			} else {
-				suite.Require().Error(err, "invalid test %d passed", i)
-			}
-		})
-	}
+func (suite *KeeperTestSuite) EvmDenom() string {
+	ctx := sdk.WrapSDKContext(suite.ctx)
+	rsp, _ := suite.queryClient.Params(ctx, &evmtypes.QueryParamsRequest{})
+	return rsp.Params.EvmDenom
 }
 
-// TestVerifyFeeAndDeductTxCostsFromUserBalance is a test method for both the VerifyFee
-// function and the DeductTxCostsFromUserBalance method.
-//
-// NOTE: This method combines testing for both functions, because these used to be
-// in one function and share a lot of the same setup.
-// In practice, the two tested functions will also be sequentially executed.
-func (suite *KeeperTestSuite) TestVerifyFeeAndDeductTxCostsFromUserBalance() {
-	hundredInt := sdkmath.NewInt(100)
-	zeroInt := sdk.ZeroInt()
-	oneInt := sdkmath.NewInt(1)
-	fiveInt := sdkmath.NewInt(5)
-	fiftyInt := sdkmath.NewInt(50)
+// Commit and begin new block
+func (suite *KeeperTestSuite) Commit() {
+	_ = suite.app.Commit()
+	header := suite.ctx.BlockHeader()
+	header.Height++
+	suite.app.BeginBlock(abci.RequestBeginBlock{
+		Header: header,
+	})
 
-	// should be enough to cover all test cases
-	initBalance := sdkmath.NewInt((ethparams.InitialBaseFee + 10) * 105)
+	// update ctx
+	suite.ctx = suite.app.NewContext(false, header)
 
-	testCases := []struct {
-		name             string
-		gasLimit         uint64
-		gasPrice         *sdkmath.Int
-		gasFeeCap        *big.Int
-		gasTipCap        *big.Int
-		cost             *sdkmath.Int
-		accessList       *ethtypes.AccessList
-		expectPassVerify bool
-		expectPassDeduct bool
-		enableFeemarket  bool
-		from             string
-		malleate         func()
-	}{
-		{
-			name:             "Enough balance",
-			gasLimit:         10,
-			gasPrice:         &oneInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "Equal balance",
-			gasLimit:         100,
-			gasPrice:         &oneInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "Higher gas limit, not enough balance",
-			gasLimit:         105,
-			gasPrice:         &oneInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: false,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "Higher gas price, enough balance",
-			gasLimit:         20,
-			gasPrice:         &fiveInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "Higher gas price, not enough balance",
-			gasLimit:         20,
-			gasPrice:         &fiftyInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: false,
-			from:             suite.address.String(),
-		},
-		// This case is expected to be true because the fees can be deducted, but the tx
-		// execution is going to fail because there is no more balance to pay the cost
-		{
-			name:             "Higher cost, enough balance",
-			gasLimit:         100,
-			gasPrice:         &oneInt,
-			cost:             &fiftyInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-		},
-		//  testcases with enableFeemarket enabled.
-		{
-			name:             "Invalid gasFeeCap w/ enableFeemarket",
-			gasLimit:         10,
-			gasFeeCap:        big.NewInt(1),
-			gasTipCap:        big.NewInt(1),
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: false,
-			expectPassDeduct: true,
-			enableFeemarket:  true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "empty tip fee is valid to deduct",
-			gasLimit:         10,
-			gasFeeCap:        big.NewInt(ethparams.InitialBaseFee),
-			gasTipCap:        big.NewInt(1),
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			enableFeemarket:  true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "effectiveTip equal to gasTipCap",
-			gasLimit:         100,
-			gasFeeCap:        big.NewInt(ethparams.InitialBaseFee + 2),
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			enableFeemarket:  true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "effectiveTip equal to (gasFeeCap - baseFee)",
-			gasLimit:         105,
-			gasFeeCap:        big.NewInt(ethparams.InitialBaseFee + 1),
-			gasTipCap:        big.NewInt(2),
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			enableFeemarket:  true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "Invalid from address",
-			gasLimit:         10,
-			gasPrice:         &oneInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: true,
-			expectPassDeduct: false,
-			from:             "abcdef",
-		},
-		{
-			name:     "Enough balance - with access list",
-			gasLimit: 10,
-			gasPrice: &oneInt,
-			cost:     &oneInt,
-			accessList: &ethtypes.AccessList{
-				ethtypes.AccessTuple{
-					Address:     suite.address,
-					StorageKeys: []common.Hash{},
-				},
-			},
-			expectPassVerify: true,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-		},
-		{
-			name:             "gasLimit < intrinsicGas during IsCheckTx",
-			gasLimit:         1,
-			gasPrice:         &oneInt,
-			cost:             &oneInt,
-			accessList:       &ethtypes.AccessList{},
-			expectPassVerify: false,
-			expectPassDeduct: true,
-			from:             suite.address.String(),
-			malleate: func() {
-				suite.ctx = suite.ctx.WithIsCheckTx(true)
-			},
-		},
+	queryHelper := baseapp.NewQueryServerTestHelper(suite.ctx, suite.app.InterfaceRegistry())
+	evmtypes.RegisterQueryServer(queryHelper, suite.app.EvmKeeper)
+	suite.queryClient = evmtypes.NewQueryClient(queryHelper)
+}
+
+func (suite *KeeperTestSuite) StateDB() *statedb.StateDB {
+	return statedb.New(suite.ctx, suite.app.EvmKeeper, statedb.NewEmptyTxConfig(common.BytesToHash(suite.ctx.HeaderHash().Bytes())))
+}
+
+// DeployTestContract deploy a test erc20 contract and returns the contract address
+func (suite *KeeperTestSuite) DeployTestContract(t require.TestingT, owner common.Address, supply *big.Int) common.Address {
+	ctx := sdk.WrapSDKContext(suite.ctx)
+	chainID := suite.app.EvmKeeper.ChainID()
+
+	ctorArgs, err := evmtypes.ERC20Contract.ABI.Pack("", owner, supply)
+	require.NoError(t, err)
+
+	nonce := suite.app.EvmKeeper.GetNonce(suite.ctx, suite.address)
+
+	data := evmtypes.ERC20Contract.Bin
+	data = append(data, ctorArgs...)
+	args, err := json.Marshal(&evmtypes.TransactionArgs{
+		From: &suite.address,
+		Data: (*hexutil.Bytes)(&data),
+	})
+	require.NoError(t, err)
+	res, err := suite.queryClient.EstimateGas(ctx, &evmtypes.EthCallRequest{
+		Args:            args,
+		GasCap:          config.DefaultGasCap,
+		ProposerAddress: suite.ctx.BlockHeader().ProposerAddress,
+	})
+	require.NoError(t, err)
+
+	var erc20DeployTx *evmtypes.MsgEthereumTx
+	if suite.enableFeemarket {
+		erc20DeployTx = evmtypes.NewTxContract(
+			chainID,
+			nonce,
+			nil,     // amount
+			res.Gas, // gasLimit
+			nil,     // gasPrice
+			suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx),
+			big.NewInt(1),
+			data,                   // input
+			&ethtypes.AccessList{}, // accesses
+		)
+	} else {
+		erc20DeployTx = evmtypes.NewTxContract(
+			chainID,
+			nonce,
+			nil,     // amount
+			res.Gas, // gasLimit
+			nil,     // gasPrice
+			nil, nil,
+			data, // input
+			nil,  // accesses
+		)
 	}
 
-	for i, tc := range testCases {
-		suite.Run(tc.name, func() {
-			suite.enableFeemarket = tc.enableFeemarket
-			suite.SetupTest()
-			vmdb := suite.StateDB()
+	erc20DeployTx.From = suite.address.Hex()
+	err = erc20DeployTx.Sign(ethtypes.LatestSignerForChainID(chainID), suite.signer)
+	require.NoError(t, err)
+	rsp, err := suite.app.EvmKeeper.EthereumTx(ctx, erc20DeployTx)
+	require.NoError(t, err)
+	require.Empty(t, rsp.VmError)
+	return crypto.CreateAddress(suite.address, nonce)
+}
 
-			if tc.malleate != nil {
-				tc.malleate()
-			}
-			var amount, gasPrice, gasFeeCap, gasTipCap *big.Int
-			if tc.cost != nil {
-				amount = tc.cost.BigInt()
-			}
+func (suite *KeeperTestSuite) TransferERC20Token(t require.TestingT, contractAddr, from, to common.Address, amount *big.Int) *evmtypes.MsgEthereumTx {
+	ctx := sdk.WrapSDKContext(suite.ctx)
+	chainID := suite.app.EvmKeeper.ChainID()
 
-			if suite.enableFeemarket {
-				if tc.gasFeeCap != nil {
-					gasFeeCap = tc.gasFeeCap
-				}
-				if tc.gasTipCap == nil {
-					gasTipCap = oneInt.BigInt()
-				} else {
-					gasTipCap = tc.gasTipCap
-				}
-				vmdb.AddBalance(suite.address, initBalance.BigInt())
-				balance := vmdb.GetBalance(suite.address)
-				suite.Require().Equal(balance, initBalance.BigInt())
-			} else {
-				if tc.gasPrice != nil {
-					gasPrice = tc.gasPrice.BigInt()
-				}
+	transferData, err := evmtypes.ERC20Contract.ABI.Pack("transfer", to, amount)
+	require.NoError(t, err)
+	args, err := json.Marshal(&evmtypes.TransactionArgs{To: &contractAddr, From: &from, Data: (*hexutil.Bytes)(&transferData)})
+	require.NoError(t, err)
+	res, err := suite.queryClient.EstimateGas(ctx, &evmtypes.EthCallRequest{
+		Args:            args,
+		GasCap:          25_000_000,
+		ProposerAddress: suite.ctx.BlockHeader().ProposerAddress,
+	})
+	require.NoError(t, err)
 
-				vmdb.AddBalance(suite.address, hundredInt.BigInt())
-				balance := vmdb.GetBalance(suite.address)
-				suite.Require().Equal(balance, hundredInt.BigInt())
-			}
-			err := vmdb.Commit()
-			suite.Require().NoError(err, "Unexpected error while committing to vmdb: %d", err)
+	nonce := suite.app.EvmKeeper.GetNonce(suite.ctx, suite.address)
 
-			tx := evmtypes.NewTx(zeroInt.BigInt(), 1, &suite.address, amount, tc.gasLimit, gasPrice, gasFeeCap, gasTipCap, nil, tc.accessList)
-			tx.From = tc.from
-
-			txData, _ := evmtypes.UnpackTxData(tx.Data)
-
-			evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
-			ethCfg := evmParams.GetChainConfig().EthereumConfig(nil)
-			baseFee := suite.app.EvmKeeper.GetBaseFee(suite.ctx, ethCfg)
-			priority := evmtypes.GetTxPriority(txData, baseFee)
-
-			fees, err := keeper.VerifyFee(txData, evmtypes.DefaultEVMDenom, baseFee, false, false, suite.ctx.IsCheckTx())
-			if tc.expectPassVerify {
-				suite.Require().NoError(err, "valid test %d failed - '%s'", i, tc.name)
-				if tc.enableFeemarket {
-					baseFee := suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx)
-					suite.Require().Equal(
-						fees,
-						sdk.NewCoins(
-							sdk.NewCoin(evmtypes.DefaultEVMDenom, sdkmath.NewIntFromBigInt(txData.EffectiveFee(baseFee))),
-						),
-						"valid test %d failed, fee value is wrong  - '%s'", i, tc.name,
-					)
-					suite.Require().Equal(int64(0), priority)
-				} else {
-					suite.Require().Equal(
-						fees,
-						sdk.NewCoins(
-							sdk.NewCoin(evmtypes.DefaultEVMDenom, tc.gasPrice.Mul(sdkmath.NewIntFromUint64(tc.gasLimit))),
-						),
-						"valid test %d failed, fee value is wrong  - '%s'", i, tc.name,
-					)
-				}
-			} else {
-				suite.Require().Error(err, "invalid test %d passed - '%s'", i, tc.name)
-				suite.Require().Nil(fees, "invalid test %d passed. fees value must be nil - '%s'", i, tc.name)
-			}
-
-			err = suite.app.EvmKeeper.DeductTxCostsFromUserBalance(suite.ctx, fees, common.HexToAddress(tx.From))
-			if tc.expectPassDeduct {
-				suite.Require().NoError(err, "valid test %d failed - '%s'", i, tc.name)
-			} else {
-				suite.Require().Error(err, "invalid test %d passed - '%s'", i, tc.name)
-			}
-		})
+	var ercTransferTx *evmtypes.MsgEthereumTx
+	if suite.enableFeemarket {
+		ercTransferTx = evmtypes.NewTx(
+			chainID,
+			nonce,
+			&contractAddr,
+			nil,
+			res.Gas,
+			nil,
+			suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx),
+			big.NewInt(1),
+			transferData,
+			&ethtypes.AccessList{}, // accesses
+		)
+	} else {
+		ercTransferTx = evmtypes.NewTx(
+			chainID,
+			nonce,
+			&contractAddr,
+			nil,
+			res.Gas,
+			nil,
+			nil, nil,
+			transferData,
+			nil,
+		)
 	}
-	suite.enableFeemarket = false // reset flag
+
+	ercTransferTx.From = suite.address.Hex()
+	err = ercTransferTx.Sign(ethtypes.LatestSignerForChainID(chainID), suite.signer)
+	require.NoError(t, err)
+	rsp, err := suite.app.EvmKeeper.EthereumTx(ctx, ercTransferTx)
+	require.NoError(t, err)
+	require.Empty(t, rsp.VmError)
+	return ercTransferTx
+}
+
+// DeployTestMessageCall deploy a test erc20 contract and returns the contract address
+func (suite *KeeperTestSuite) DeployTestMessageCall(t require.TestingT) common.Address {
+	ctx := sdk.WrapSDKContext(suite.ctx)
+	chainID := suite.app.EvmKeeper.ChainID()
+
+	data := evmtypes.TestMessageCall.Bin
+	args, err := json.Marshal(&evmtypes.TransactionArgs{
+		From: &suite.address,
+		Data: (*hexutil.Bytes)(&data),
+	})
+	require.NoError(t, err)
+
+	res, err := suite.queryClient.EstimateGas(ctx, &evmtypes.EthCallRequest{
+		Args:            args,
+		GasCap:          config.DefaultGasCap,
+		ProposerAddress: suite.ctx.BlockHeader().ProposerAddress,
+	})
+	require.NoError(t, err)
+
+	nonce := suite.app.EvmKeeper.GetNonce(suite.ctx, suite.address)
+
+	var erc20DeployTx *evmtypes.MsgEthereumTx
+	if suite.enableFeemarket {
+		erc20DeployTx = evmtypes.NewTxContract(
+			chainID,
+			nonce,
+			nil,     // amount
+			res.Gas, // gasLimit
+			nil,     // gasPrice
+			suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx),
+			big.NewInt(1),
+			data,                   // input
+			&ethtypes.AccessList{}, // accesses
+		)
+	} else {
+		erc20DeployTx = evmtypes.NewTxContract(
+			chainID,
+			nonce,
+			nil,     // amount
+			res.Gas, // gasLimit
+			nil,     // gasPrice
+			nil, nil,
+			data, // input
+			nil,  // accesses
+		)
+	}
+
+	erc20DeployTx.From = suite.address.Hex()
+	err = erc20DeployTx.Sign(ethtypes.LatestSignerForChainID(chainID), suite.signer)
+	require.NoError(t, err)
+	rsp, err := suite.app.EvmKeeper.EthereumTx(ctx, erc20DeployTx)
+	require.NoError(t, err)
+	require.Empty(t, rsp.VmError)
+	return crypto.CreateAddress(suite.address, nonce)
 }
